refactor: LessonRequest 가격관련 필드 Int -> Long으로 수정

+++ src/main/kotlin/team/mozu/dsm/adapter/in/lesson/dto/request/LessonRequest.kt
@@ -15,7 +15,7 @@ data class LessonRequest(
 
     @field:Min(100_000, message = "기초 자산은 최소 100,000원 이상 입력해야 합니다.")
     @field:Max(2_147_483_647, message = "기초 자산은 2,147,483,647원 이하로 입력해야 합니다.")
-    val baseMoney: Int,
+    val baseMoney: Long,
 
     @field:Min(3, message = "투자 차수는 최소 3차부터 가능합니다.")
     @field:Max(5, message = "투자 차수는 최대 5차까지 가능합니다.")
@@ -31,7 +31,7 @@ data class LessonRequest(
 
 data class LessonItemRequest(
     val itemId: Int,
-    val money: List<Int>
+    val money: List<Long>
 )
 
 data class LessonArticleRequest(
